calculator/requirements.txt
click              8.0.1
Flask              2.0.1
grpcio             1.39.0
grpcio-tools       1.39.0
importlib-metadata 4.6.4
itsdangerous       2.0.1
Jinja2             3.0.1
MarkupSafe         2.0.1
pip                21.2.3
protobuf           3.17.3
setuptools         57.4.0
six                1.16.0
typing-extensions  3.10.0.0
Werkzeug           2.0.1
wheel              0.37.0
zipp               3.5.0
